Moves DiceAmountCalculator to combat package
