+++ core/src/main/combat/ClassicCombatDiceAmountCalculator.kt
@@ -1,4 +1,4 @@
-package situations
+package combat
 
 import PositiveInt as Pos
 
@@ -6,7 +6,8 @@ class ClassicCombatDiceAmountCalculator(
     private val attackingArmies: Pos, private val defendingArmies: Pos) {
 
     fun forAttacker(): Pos {
-        var dice = minOf(MAXIMUM_DICE_AMOUNT, attackingArmies - Pos(1))
+        var dice = minOf(
+            MAXIMUM_DICE_AMOUNT, attackingArmies - Pos(1))
         if (shouldAddOneDieToAttacker()) {
             dice++
         }
@@ -16,7 +17,8 @@ class ClassicCombatDiceAmountCalculator(
     private fun shouldAddOneDieToAttacker() =
         attackingArmies >= Pos(2) * defendingArmies && defendingArmies >= Pos(3)
 
-    fun forDefender() = minOf(MAXIMUM_DICE_AMOUNT, defendingArmies)
+    fun forDefender() = minOf(
+        MAXIMUM_DICE_AMOUNT, defendingArmies)
 
     companion object {
         private val MAXIMUM_DICE_AMOUNT = Pos(3)
